third_party/WebKit/LayoutTests/fast/block/large-border-abspos.html

<!DOCTYPE html>
<!-- The left offset here is just to push the huge thing out of view, so that a
     human (yes, YOU!) can see the test results. -->
<div style="position:absolute; left:100%; border-width:1234567890px; border-style:solid;">
    <div id="abspos" style="position:absolute; left:0; top:0; right:0; bottom:0;"></div>
</div>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
    test(() => {
        assert_equals(document.getElementById("abspos").offsetWidth, 0);
        assert_equals(document.getElementById("abspos").offsetHeight, 0);
    }, "Sizes should never become negative");
</script>
